Resolvendo codacy

+++ aplicacao/contest/src/main/java/ufc/quixada/npi/contest/service/EventoService.java
@@ -1,6 +1,7 @@
 package ufc.quixada.npi.contest.service;
 
 import java.util.List;
+import java.util.logging.Logger;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,6 +21,7 @@ import ufc.quixada.npi.contest.repository.EventoRepository;
 import ufc.quixada.npi.contest.util.Constants;
 import ufc.quixada.npi.contest.util.ContestUtil;
 
+
 @Service
 public class EventoService {
 
@@ -27,7 +29,7 @@ public class EventoService {
 	private static final String TITULO_EMAIL_ORGANIZADOR = "TEXTO_EMAIL_CONVITE_ORGANIZADOR";
 	private static final String ASSUNTO_EMAIL_CONFIRMACAO = "ASSUNTO_EMAIL_CONFIRMACAO";
 	private static final String TEXTO_EMAIL_CONFIRMACAO = ".Fique atento aos prazos, o próximo passo será a fase das revisões, confira no edital os prazos. Boa sorte!";
-	
+
 	@Autowired
 	EventoRepository eventoRepository;
 
@@ -42,13 +44,13 @@ public class EventoService {
 
 	@Autowired
 	ParticipacaoEventoService participacaoEventoService;
-	
+
 	@Autowired
 	private TokenService tokenService;
-	
+
 	@Autowired
 	private TrilhaService trilhaService;
-	
+
 	@Autowired
 	private UsuarioService usuarioService;
 
@@ -58,39 +60,40 @@ public class EventoService {
 		String nome = "Nome Temporário";
 
 		String assunto = messageService.getMessage(TITULO_EMAIL_ORGANIZADOR) + " " + evento.getNome();
-		String corpo = "Olá"+ messageService.getMessage(TEXTO_EMAIL_ORGANIZADOR) + " " + evento.getNome() + " como "+ papel.getNome();
+		String corpo = "Olá" + messageService.getMessage(TEXTO_EMAIL_ORGANIZADOR) + " " + evento.getNome() + " como "
+				+ papel.getNome();
 		String titulo = "[CONTEST] Convite para o Evento: " + evento.getNome();
 		String pageCadastro = "/completar-cadastro/";
-		Token token =  new Token();
-		
+		Token token = new Token();
+
 		if (pessoa == null) {
-			
+
 			try {
 				Usuario usuarioLdap = usuarioService.getByEmail(email);
-				
-				if(usuarioLdap != null){
+
+				if (usuarioLdap != null) {
 					pessoa = ContestUtil.convertUsuarioToPessoa("", usuarioLdap);
-				}else{
+				} else {
 					pessoa = new Pessoa(nome, email);
 					pessoa.setPapel(Tipo.USER);
 				}
 				pessoaService.addOrUpdate(pessoa);
-			
+
 				token = tokenService.novoToken(pessoa, Constants.ACAO_COMPLETAR_CADASTRO);
 			} catch (Exception e) {
 				e.printStackTrace();
 			}
-			
-			corpo = corpo + ". Você não está cadastrado na nossa base de dados. Acesse: " + url + pageCadastro + token.getToken() + " e termine o seu cadastro";
+
+			corpo = corpo + ". Você não está cadastrado na nossa base de dados. Acesse: " + url + pageCadastro
+					+ token.getToken() + " e termine o seu cadastro";
 		} else {
 			corpo = corpo + ". Realize o login em " + url + "/login";
 		}
-		
-		
+
 		if (!emailService.enviarEmail(titulo, assunto, email, corpo)) {
-			try{
+			try {
 				pessoaService.delete(pessoa.getId());
-			}catch (Exception ex){
+			} catch (Exception ex) {
 				ex.printStackTrace();
 			}
 			return false;
@@ -114,42 +117,42 @@ public class EventoService {
 		Tipo papel = Tipo.AUTOR;
 		return adicionarPessoa(email, evento, papel, url);
 	}
-	
+
 	private boolean notificarPessoa(String email, Evento evento, Tipo papel, Trabalho trabalho) {
 
 		Pessoa pessoa = pessoaService.getByEmail(email);
 		String nome = "Nome Temporário";
 
 		String assunto = messageService.getMessage(ASSUNTO_EMAIL_CONFIRMACAO) + " " + trabalho.getTitulo();
-		String corpo = "Olá, seu trabalho " + trabalho.getTitulo() + " foi enviado com sucesso para o evento " + evento.getNome() + TEXTO_EMAIL_CONFIRMACAO;
+		String corpo = "Olá, seu trabalho " + trabalho.getTitulo() + " foi enviado com sucesso para o evento "
+				+ evento.getNome() + TEXTO_EMAIL_CONFIRMACAO;
 		String titulo = "[CONTEST] Confirmação de envio do trabalho: " + trabalho.getTitulo();
-		
-		
+
 		if (pessoa == null) {
-			
+
 			try {
 				Usuario usuarioLdap = usuarioService.getByEmail(email);
-				
-				if(usuarioLdap != null){
+
+				if (usuarioLdap != null) {
 					pessoa = ContestUtil.convertUsuarioToPessoa("", usuarioLdap);
-				}else{
+				} else {
 					pessoa = new Pessoa(nome, email);
 					pessoa.setPapel(Tipo.USER);
 				}
 				pessoaService.addOrUpdate(pessoa);
-			
+
 			} catch (Exception e) {
-				e.printStackTrace();
+				final Logger lOGGER = Logger.getLogger("");
+				lOGGER.log(java.util.logging.Level.INFO, "", e);
 			}
 		}
-			
-		
-		
+
 		if (!emailService.enviarEmail(titulo, assunto, email, corpo)) {
-			try{
-				pessoaService.delete(pessoa.getId());
-			}catch (Exception ex){
-				ex.printStackTrace();
+			try {
+				if(!(pessoa == null)) pessoaService.delete(pessoa.getId());
+			} catch (Exception ex) {
+				final Logger lOGGER = Logger.getLogger("");
+				lOGGER.log(java.util.logging.Level.INFO, "", ex);
 			}
 			return false;
 		}
@@ -157,17 +160,17 @@ public class EventoService {
 		participacaoEventoService.adicionarOuEditarParticipacaoEvento(participacao);
 		return true;
 	}
-	
+
 	public boolean notificarAutor(String email, Evento evento, Trabalho trabalho) {
 		Tipo papel = Tipo.AUTOR;
-		return notificarPessoa(email, evento, papel,trabalho);
+		return notificarPessoa(email, evento, papel, trabalho);
 	}
 
-	public boolean notificarCoautor(String email, Evento evento,Trabalho trabalho) {
+	public boolean notificarCoautor(String email, Evento evento, Trabalho trabalho) {
 		Tipo papel = Tipo.COAUTOR;
-		return notificarPessoa(email, evento, papel,trabalho);
+		return notificarPessoa(email, evento, papel, trabalho);
 	}
-	
+
 	public boolean adicionarOuAtualizarEvento(Evento evento) {
 		if (evento.getPrazoSubmissaoInicial() != null && evento.getPrazoSubmissaoFinal() != null
 				&& evento.getPrazoRevisaoInicial() != null && evento.getPrazoRevisaoFinal() != null) {
@@ -176,7 +179,7 @@ public class EventoService {
 					&& evento.getPrazoRevisaoInicial().after(evento.getPrazoSubmissaoInicial())
 					&& evento.getPrazoRevisaoInicial().before(evento.getPrazoRevisaoFinal())
 					&& evento.getPrazoRevisaoFinal().before(evento.getPrazoSubmissaoFinal())) {
-				
+
 				eventoRepository.save(evento);
 				return true;
 			}
@@ -252,13 +255,14 @@ public class EventoService {
 		return eventoRepository.findEventoByParticipacoesPessoaIdAndParticipacoesPapelAndEstado(idOrganizador,
 				Tipo.ORGANIZADOR, EstadoEvento.INATIVO);
 	}
+
 	public List<Evento> buscarEventosQueReviso(Long idRevisor) {
-		return eventoRepository.findEventoByParticipacoesPessoaIdAndParticipacoesPapelAndEstado(idRevisor, Tipo.REVISOR, EstadoEvento.ATIVO);
+		return eventoRepository.findEventoByParticipacoesPessoaIdAndParticipacoesPapelAndEstado(idRevisor, Tipo.REVISOR,
+				EstadoEvento.ATIVO);
 	}
 
 	public List<Evento> getEventosByEstadoEVisibilidadePublica(EstadoEvento estado) {
 		return eventoRepository.findEventoByEstadoAndVisibilidade(estado, VisibilidadeEvento.PUBLICO);
 	}
-
 
 }
